reorganize to fix linting
