+++ integration/tests/access/cohort3/access_circuit_breaker_test.go
@@ -1,4 +1,4 @@
-package cohort4
+package cohort3
 
 import (
 	"context"
